Updated variables

+++ Assets/Scripts/GameManager.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour {
 
-	public float lives = 3;
-	public float coinChance = 0.3;
-	public float growPaddleChance = 0.2;
-	public float extraLife = 0.05;
-	public float multipleBall = 0.1;
-	public float shrinkPowerDown = 0.3;
-	public float stickyPower = 0.4;
-	public float paddleUp = 0.25;
-	public float starPower = 0.02;
-	public float super = 0.01;
+	public double lives = 3;
+	public double coinChance = 0.3;
+	public double growPaddleChance = 0.2;
+	public double extraLife = 0.05;
+	public double multipleBall = 0.1;
+	public double shrinkPowerDown = 0.3;
+	public double stickyPower = 0.4;
+	public double paddleUp = 0.25;
+	public double starPower = 0.02;
+	public double super = 0.01;
 	// Use this for initialization
 	void Start () {
